+++ examples/equalizer-filter/src/lib.rs
@@ -124,7 +124,7 @@ impl aviutl2::filter::FilterPlugin for EqualizerFilter {
             ),
             flags: aviutl2::bitflag!(aviutl2::filter::FilterPluginFlags {
                 audio: true,
-                as_filter: true,
+                filter: true,
             }),
             config_items: FilterConfig::to_config_items(),
         }